# back/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Strapp Template API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API v1"
        }
    ],
    "paths": {
...
        "/admin/files/upload": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Upload a file to S3 as admin",
                "operationId": "06403a144f07808091fe1a64a2f5c264",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "file"
                                ],
                                "properties": {
                                    "file": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Uploaded"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/auth/user/verify-2fa": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Verify 2FA OTP code and receive access token",
                "operationId": "196e3fd197825a4a5729b3678194686e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "session_token",
                                    "code"
                                ],
                                "properties": {
                                    "session_token": {
                                        "type": "string",
                                        "example": "uuid-string"
                                    },
                                    "code": {
                                        "type": "string",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Verification success — token + user returned",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "token": {
                                            "type": "string"
                                        },
                                        "token_type": {
                                            "type": "string",
                                            "example": "Bearer"
                                        },
                                        "user": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Invalid code, expired session, or invalid session token"
                    },
                    "429": {
                        "description": "Session locked due to too many attempts"
                    }
                }
            }
        },
        "/api/v1/auth/user/resend-2fa": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Resend 2FA OTP code (60s cooldown)",
                "operationId": "ed244092f4f870203126bcd07c3de1f9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "session_token"
                                ],
                                "properties": {
                                    "session_token": {
                                        "type": "string",
                                        "example": "uuid-string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Code resent successfully"
                    },
                    "422": {
                        "description": "Invalid or expired session"
                    },
                    "429": {
                        "description": "Cooldown not elapsed yet"
                    }
                }
            }
        },
        "/api/v1/user/two-factor": {
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "Toggle two-factor authentication on or off",
                "operationId": "80c1399a3aebb4a320a04988b7c7635e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "enabled",
                                    "password"
                                ],
                                "properties": {
                                    "enabled": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "2FA toggled successfully"
                    },
                    "422": {
                        "description": "Invalid password"
                    },
                    "403": {
                        "description": "Forbidden — missing permission"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}